+++ R/group_polys.R
@@ -13,7 +13,7 @@
 #' ## R-spatial evolution
 #'
 #' Please note, spatsoc has followed updates from R spatial, GDAL and PROJ for
-#' handling projections, see more below and  details at
+#' handling coordinate reference systems, see more below and  details at
 #' <https://r-spatial.org/r/2020/03/17/wkt.html>.
 #'
 #' In addition, `group_polys` previously used rgeos::gIntersection,
@@ -32,12 +32,10 @@
 #' the names of respective columns in `DT` which correspond to the
 #' individual identifier, X and Y coordinates, and additional grouping columns.
 #'
-#' The `projection` argument expects a character string or numeric
-#' defining the coordinate reference system to be passed to [sf::st_crs].
-#' For example, for UTM zone 36S (EPSG 32736), the projection
-#' argument is `projection = "EPSG:32736"` or `projection = 32736`.
-#' See <https://spatialreference.org>
-#' for a list of EPSG codes.
+#' The `crs` argument expects a character string or numeric defining the
+#' coordinate reference system to be passed to [sf::st_crs]. For example, for
+#' UTM zone 36S (EPSG 32736), the crs argument is `crs = "EPSG:32736"` or `crs =
+#' 32736`. See <https://spatialreference.org> for a list of EPSG codes.
 #'
 #' The `hrType` must be either one of "kernel" or "mcp". The
 #' `hrParams` must be a named list of arguments matching those of
@@ -101,11 +99,11 @@
 #' utm <- 32736
 #'
 #' group_polys(DT, area = FALSE, hrType = 'mcp',
-#'             hrParams = list(percent = 95), projection = utm,
+#'             hrParams = list(percent = 95), crs = utm,
 #'             id = 'ID', coords = c('X', 'Y'))
 #'
 #' areaDT <- group_polys(DT, area = TRUE, hrType = 'mcp',
-#'                       hrParams = list(percent = 95), projection = utm,
+#'                       hrParams = list(percent = 95), crs = utm,
 #'                       id = 'ID', coords = c('X', 'Y'))
 #' print(areaDT)
 group_polys <-
@@ -113,13 +111,19 @@ group_polys <-
            area = NULL,
            hrType = NULL,
            hrParams = NULL,
-           projection = NULL,
+           crs = NULL,
            id = NULL,
            coords = NULL,
            splitBy = NULL,
-           sfPolys = NULL) {
+           sfPolys = NULL,
+           projection = NULL) {
     # due to NSE notes in R CMD check
     nBy <- ..coords <- ..id <- withinGroup <- group <- outGroup <- NULL
+
+    if (!is.null(projection)) {
+      warning('projection argument is deprecated, setting crs = projection')
+      crs <- projection
+    }
 
     if (is.null(area) || !is.logical(area)) {
       stop('area must be provided (TRUE or FALSE)')
@@ -138,7 +142,7 @@ group_polys <-
         sfPolys <-
           build_polys(
             DT = DT,
-            projection = projection,
+            crs = crs,
             hrType = hrType,
             hrParams = hrParams,
             coords = coords,
@@ -236,7 +240,7 @@ group_polys <-
               sfPolys <-
                 build_polys(
                   DT = .SD,
-                  projection = projection,
+                  crs = crs,
                   hrType = hrType,
                   hrParams = hrParams,
                   coords = ..coords,
@@ -285,7 +289,7 @@ group_polys <-
               sfPolys <-
                 build_polys(
                   DT = .SD,
-                  projection = projection,
+                  crs = crs,
                   hrType = hrType,
                   hrParams = hrParams,
                   id = ..id,