+++ tests/testthat/test-direction-polarization.R
@@ -11,13 +11,13 @@ threshold <- 50
 coords <- c('X', 'Y')
 timegroup <- 'timegroup'
 group <- 'group'
-projection <- 32736
+utm <- 32736
 
 DT[, datetime := as.POSIXct(datetime, tz = 'UTC')]
 group_times(DT, datetime = datetime, threshold = timethreshold)
 group_pts(DT, threshold = threshold, id = id,
           coords = coords, timegroup = timegroup)
-direction_step(DT, id = id, coords = coords, projection = projection)
+direction_step(DT, id = id, coords = coords, crs = utm)
 
 clean_DT <- copy(DT)