+++ tests/testthat/test-edge-alignment.R
@@ -8,7 +8,7 @@ coords <- c('X', 'Y')
 direction <- 'direction'
 timegroup <- 'timegroup'
 group <- 'group'
-projection <- 32736
+utm <- 32736
 
 DT <- fread('../testdata/DT.csv')
 group_times(DT, 'datetime', '10 minutes')
@@ -16,7 +16,7 @@ direction_step(
   DT = DT,
   id = id,
   coords = coords,
-  projection = projection
+  crs = utm
 )
 
 test_that('required arguments are provided else error', {